ubn_requirements.txt
amqp==2.2.2
asn1crypto==0.24.0
bcrypt==3.1.4
billiard==3.5.0.3
celery==4.1.0
cffi==1.11.5
cryptography==2.2.2
distro==1.2.0
Django==2.0.4
django-appconf==1.0.2
django-compressor==2.2
django-datatables-view==1.16.0
django-discover-runner==1.0
django-filter==1.1.0
django-role-permissions==2.2.0
djangorestframework==3.8.2
drfdocs==0.0.11
idna==2.6
kombu==4.1.0
ldap3==2.5
paramiko==2.4.1
pexpect==4.5.0
ptyprocess==0.5.2
pyasn1==0.4.2
pycparser==2.18
Pygments==2.2.0
PyNaCl==1.2.1
pytz==2018.4
rcssmin==1.0.6
rjsmin==1.0.12
scp==0.10.2
six==1.11.0
vine==1.1.4
